// apps/ui-admin/src/components/layouts/app/pages/my-listings/components/my-listings-dashboard.types.ts
export interface MyListingData {
  id: string;
  title: string;
  image?: string | null;
  createdAt?: string | null;
  reservationPeriod?: string | null;
  status: string;
  pendingRequestsCount: number;
}

export interface ListingRequestData {
  id: string;
  title: string;
  image?: string | null;
  requestedBy: string;
  requestedOn: string;
  reservationPeriod: string;
  status: string;
}